Multiple aliases, negates, and POSIX Style
Multiple aliases allow to specify a list of aliases in the alias
argument.

Same for the negate argument.

Also, added a POSIX feature that may be enabled at the Command or Group
level.

If a non-bool alias starts with a dash, then a space may replace the
equal: `-a=b` is same as `-a b`