New cli2.Model for cli2.Client
So, in a gig a couple of years age I got my kicks out of implementing a
Django-like framework to create a Python SDK for a proprietary API.

And I keep getting this kind of gigs, typical from big-corp, where you
are given an API of some proprietary product, with not even a passable
CLI to consume it if any.

Basically, I'm open sourcing that, because I ended coding the same for
another gig.

There's a lot more features to come, namely descriptors, although, this
commit succeeds in encapsulating the complexity of creating an HTTP
Client library with a generated CLI.